+++ ruma-common/CHANGELOG.md
@@ -4,6 +4,10 @@ Breaking changes:
 
 * Update strum dependency to 0.19
 
+Improvements:
+
+* Add the `thirdparty` module
+
 # 0.2.0
 
 Breaking changes: