+++ ruma-common/CHANGELOG.md
@@ -9,6 +9,7 @@ Improvements:
 * Add the `thirdparty` module
 * Add `directory::{Filter, PublicRoomsChunk, RoomNetwork}` (moved from
   `ruma_client_api::r0::directory`)
+* Add `push::{PusherData, PushFormat}` (moved from `ruma_client_api::r0::push`)
 
 # 0.2.0